requirements.txt
pydub
openai-whisper
librosa
soundfile
Flask
